requirements.txt
kivy
numpy
pandas
Pillow
scipy
pyttsx3
